refactor(kernel/declaration): remove `self_opt` flag from reducibility hints
This flag was used by the kernel to decide whether the following
heuristic should be used to avoid unfolding `f` at `is_def_eq`.

       f a =?= f b
       -----------
         a =?= b

This heuristic was introduced at Lean1 after a discussion with
Georges Gontier. Since this discussion, we added support for
caching failures of this heuristic. This proved to be much more
effective to attack the performance problems.
Moreover, we do not even use this flag in the `type_context::is_def_eq`
used during elaboration.

The current codebase contains only one place where this flag was set to
`false`: coercions generated at structure_cmd. This change was
made at commit
https://github.com/leanprover/lean4/commit/1c70514231ebc51836bc4e558869ebed57559e8e
in the Lean2 codebase when we were not caching failures and
the kernel type checker was also used during elaboration.